% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sundial.R
\name{create_sundial}
\alias{create_sundial}
\title{Create an analemmatic sundial}
\usage{
create_sundial(latitude, magnetic_declination)
}
\arguments{
\item{latitude}{degrees}

\item{magnetic_declination}{degrees. Positive is EAST.}
}
\value{
ggplot object
}
\description{
Create an analemmatic sundial
}
