% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sundial.R
\name{create_sundial_for_city}
\alias{create_sundial_for_city}
\title{Create an analemmatic sundial for a given city}
\usage{
create_sundial_for_city(city, country = NULL, ...)
}
\arguments{
\item{city}{city name}

\item{country}{default: NULL}

\item{...}{arguments passed to \code{create_sundial}}
}
\value{
ggplot object
}
\description{
Create an analemmatic sundial for a given city
}
\details{
Create an analemmatic sundial for a given city.  The country name may be
specified, otherwise the first country it is within is chosen.  See
\code{analemmatic:::cities} for the raw city information.

The Lat/Long for \code{cities} is from \code{maps::world.cities}.  The
magnetic declination has been calculated offline for each city using
\code{oce::magneticField()}.
}
